pyproject.toml
[project]
name = "eternal-memory"
version = "0.1.0"
description = "OpenClaw-style Eternal Memory System - A local-first AI memory engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "openai>=1.12.0",
    "pyyaml>=6.0.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eternal_memory"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
